requirements-gpu.txt
# GPU Training Requirements - CUDA 12.9
# 
# Installation instructions:
# 1. Install PyTorch with CUDA 12.9:
#    pip install torch==2.8.0+cu129 torchvision==0.23.0+cu129 --index-url https://download.pytorch.org/whl/cu129
# 2. Install other packages:
#    pip install ultralytics opencv-python Pillow numpy pandas matplotlib tqdm PyYAML

# Computer Vision and ML frameworks
ultralytics>=8.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data handling and utilities
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0

# Optional: Additional CUDA utilities (uncomment if needed)
# torchaudio==2.8.0+cu129